feature(jwt): adding token table and implementing user handler

+++ internal/app/user_service.go
@@ -30,14 +30,44 @@ func (s *UserService) CreateUser(ctx context.Context, user *domain.User) (string
 	return s.userRepo.Save(ctx, user)
 }
 
-func (s *UserService) AddContact(ctx context.Context, userID, contactID string) error {
+func (s *UserService) Login(ctx context.Context, username, password string) (*domain.RegisterResponse, error) {
+
+	return nil, nil
+}
+
+func (s *UserService) RegisterUser(ctx context.Context, user *domain.User) (*domain.RegisterResponse, error) {
+
+	return nil, nil
+}
+
+func (s *UserService) RefreshToken(ctx context.Context, refreshToken string) (*domain.RefreshResponse, error) {
+
+	return nil, nil
+}
+
+func (s *UserService) RevokeToken(ctx context.Context, refreshToken string) error {
+
+	return nil
+}
+
+func (s *UserService) AddContact(ctx context.Context, contactID string) error {
+	claims, ok := s.authService.ClaimsFromContext(ctx)
+	if !ok {
+		return domain.ErrUnauthorized
+	}
+	userID := claims.UserID
 	if err := s.userRepo.AddContact(ctx, userID, contactID); err != nil {
 		return err
 	}
 	return nil
 }
 
-func (s *UserService) RemoveContact(ctx context.Context, userID, contactID string) error {
+func (s *UserService) RemoveContact(ctx context.Context, contactID string) error {
+	claims, ok := s.authService.ClaimsFromContext(ctx)
+	if !ok {
+		return domain.ErrUnauthorized
+	}
+	userID := claims.UserID
 	if err := s.userRepo.RemoveContact(ctx, userID, contactID); err != nil {
 		return err
 	}
